% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{format.informative_na}
\alias{format.informative_na}
\alias{vec_ptype_abbr.informative_na}
\alias{vec_ptype_full.informative_na}
\title{Format method for informative_na}
\usage{
\method{format}{informative_na}(x, ...)

\method{vec_ptype_abbr}{informative_na}(x, ...)

\method{vec_ptype_full}{informative_na}(x, ...)
}
\arguments{
\item{x}{A informative_na vector.}

\item{...}{Additional arguments passed to \code{\link[=format]{format()}}.}
}
\description{
Format method for informative_na
}
\keyword{internal}
