% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{missing_predicates}
\alias{missing_predicates}
\alias{is_observed}
\alias{is_mcar}
\alias{is_mar}
\alias{is_mnar}
\alias{is_ignorable}
\title{Extract missingness predicates}
\usage{
is_observed(x)

is_mcar(x)

is_mar(x)

is_mnar(x)

is_ignorable(x)
}
\arguments{
\item{x}{A informative_na vector.}
}
\value{
A logical vector of the same length as x.
}
\description{
These functions extract logical masks indicating which values have a specific
missingness type (MCAR, MAR, or MNAR).
}
\examples{
x <- why_na(c(1.5, NA, 2.3, NA), c(0L, 1L, 0L, 3L))
is_mcar(x)   # TRUE where type == 1L
is_mar(x)    # TRUE where type == 2L
is_mnar(x)   # TRUE where type == 3L

}
