% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{missing_reason}
\alias{missing_reason}
\title{Return missingness type as a factor}
\usage{
missing_reason(x)
}
\arguments{
\item{x}{A informative_na vector.}
}
\value{
A factor with levels "Observed", "MCAR", "MAR", "MNAR".
}
\description{
Return missingness type as a factor
}
\examples{
x <- why_na(c(1.5, NA, 2.3, NA), c(0L, 1L, 0L, 3L))
missing_reason(x)

}
