% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arch.R
\name{new_informative_na}
\alias{new_informative_na}
\alias{why_na}
\alias{informative_na}
\title{Create an informative NA vector}
\usage{
new_informative_na(value = double(), type = integer())

why_na(x, type = NULL)
}
\arguments{
\item{value}{A double vector of values. Will be coerced to double.}

\item{type}{An integer vector indicating missingness type.
0L = Observed, 1L = MCAR, 2L = MAR, 3L = MNAR.}

\item{x}{A vector to coerce to double (or NULL).}
}
\value{
A vctrs_rcrd of class informative_na.
}
\description{
Low-level constructor that creates a vctrs record with value and type fields.
Enforces immutability: values with type != 0L are forced to NA_real_.
}
\keyword{internal}
