% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whyna-package.R
\docType{package}
\name{whyna-package}
\alias{whyna}
\alias{whyna-package}
\title{whyna: Informative Missingness via vctrs Records}
\description{
Implements typed/informative missingness following Rubin's MCAR/MAR/MNAR framework. Tracks why values are missing using a vctrs record vector with 'value' and 'type' fields. Integrates cleanly with tibble, dplyr, and base R idioms.
}
\author{
\strong{Maintainer}: Andrew Brown \email{brown.andrewg@gmail.com} (\href{https://orcid.org/0000-0002-4565-533X}{ORCID})

}
\keyword{internal}
